backend/requirements.txt
fastapi
uvicorn
pandas
numpy
scipy
yfinance
openpyxl
SQLAlchemy
python-jose[cryptography]
python-multipart
python-dotenv

matplotlib
seaborn
httpx
bcrypt